Refactorment for Navigation components. Moved them to a seperate folder
