+++ src/App.vue
@@ -4,7 +4,7 @@
 </template>
 
 <script>
-import MainNav from "@/components/MainNav.vue";
+import MainNav from "@/components/Navigation/MainNav.vue";
 import MainHero from "@/components/MainHero.vue";
 
 export default {
