+++ src/components/Navigation/MainNav.vue
@@ -35,8 +35,8 @@
 
 <script>
 import ActionButton from "@/components/ActionButton.vue";
-import ProfileImage from "@/components/ProfileImage.vue";
-import SubNav from "@/components/SubNav.vue";
+import ProfileImage from "@/components/Navigation/ProfileImage.vue";
+import SubNav from "@/components/Navigation/SubNav.vue";
 
 export default {
   name: "MainNav",
