+++ src/components/Navigation/ProfileImage.vue
