+++ tests/unit/components/Navigation/SubNav.test.js
@@ -1,5 +1,5 @@
 import { render, screen } from "@testing-library/vue";
-import SubNav from "@/components/SubNav.vue";
+import SubNav from "@/components/Navigation/SubNav.vue";
 
 describe("SubNav", () => {
   const renderSubNav = (onJobsResultPage) => {
